# .docs/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    url: https://github.com/Payback159/tenama/issues/new
  description: "Tenama provides a simple REST API to enable non-cluster-admins to\
    \ create temporary namespaces. tenama takes care of the creation, management\
    \ and cleanup of the temporary namespaces."
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Tenama
  version: 0.1.0
servers:
- url: /
tags:
- externalDocs:
    description: Find out more
    url: https://github.com/Payback159/tenama
  name: Documentation
- description: Everything about your temporarily namespaces
  name: Namespaces
paths:
  /docs:
    get:
      operationId: getDocumentation
      responses:
        "200":
          content:
            text/html:
              schema:
                example: |-
                  <html>
                  <head>
                  </head>
                  <body>
                  </body>
                  </html>
                type: string
          description: successful operation
      summary: Outputs the openAPI specification
      tags:
      - Documentation
  /namespace:
    post:
      operationId: createNamespace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Namespace'
        description: Namespace object that needs to be added to the cluster
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: string
        "401":
          description: Authentication information is missing or invalid
          headers:
            WWW_Authenticate:
              explode: false
              schema:
                type: string
              style: simple
        "405":
          description: Invalid input
      security:
      - basicAuth: []
      summary: Create a new namespace
      tags:
      - Namespaces
      x-codegen-request-body-name: body
  /namespace/{namespace}:
    delete:
      operationId: deleteNamespace
      parameters:
      - description: Namespace to delete
        explode: false
        in: path
        name: namespace
        required: true
        schema:
          type: string
        style: simple
      responses:
        "400":
          content: {}
          description: Invalid Namespace supplied
        "401":
          description: Authentication information is missing or invalid
          headers:
            WWW_Authenticate:
              explode: false
              schema:
                type: string
              style: simple
        "404":
          content: {}
          description: Namespace not found
      security:
      - basicAuth: []
      summary: Deletes a namespace
      tags:
      - Namespaces
    get:
      description: Returns a single namespace
      operationId: getNamespaceByName
      parameters:
      - description: name of namespace to return
        explode: false
        in: path
        name: namespace
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Namespace'
          description: successful operation
        "400":
          content: {}
          description: Invalid namespace supplied
        "401":
          description: Authentication information is missing or invalid
          headers:
            WWW_Authenticate:
              explode: false
              schema:
                type: string
              style: simple
        "404":
          content: {}
          description: Namespace not found
      security:
      - basicAuth: []
      summary: Find namespace by name
      tags:
      - Namespaces
components:
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      headers:
        WWW_Authenticate:
          explode: false
          schema:
            type: string
          style: simple
  schemas:
    Namespace:
      example:
        duration: duration
        infix: infix
        suffix: suffix
        users: users
      properties:
        infix:
          type: string
        suffix:
          type: string
        duration:
          description: How long should the namespace be preserved until it becomes
            obsolete and is automatically cleaned up.
          type: string
        users:
          description: A list of users to be authorized as editors in this namespace.
          items:
            type: string
          type: array
      type: object
      xml:
        name: Namespace
  securitySchemes:
    basicAuth:
      scheme: basic
      type: http